requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
httpx>=0.24.1

# LangChain and LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langchain-community>=0.0.13
langchain-groq>=0.2.5
langgraph-supervisor>=0.0.9

# For API Calls
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1

# Environment variables
python-dotenv>=1.0.0
